# Caching parsed expressions in sql planner