--- a/processing/src/main/java/org/apache/druid/query/JoinDataSource.java
+++ b/processing/src/main/java/org/apache/druid/query/JoinDataSource.java
@@ -24,11 +24,13 @@ import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import com.google.common.base.Preconditions;
 import com.google.common.base.Supplier;
+import com.google.common.base.Suppliers;
 import com.google.common.collect.ImmutableList;
 import org.apache.druid.java.util.common.IAE;
 import org.apache.druid.java.util.common.StringUtils;
 import org.apache.druid.math.expr.Expr;
 import org.apache.druid.math.expr.ExprMacroTable;
+import org.apache.druid.math.expr.Parser;
 import org.apache.druid.query.filter.DimFilter;
 import org.apache.druid.segment.join.JoinConditionAnalysis;
 import org.apache.druid.segment.join.JoinPrefixUtils;
@@ -102,17 +104,14 @@ public class JoinDataSource implements DataSource
       @JacksonInject ExprMacroTable macroTable
   )
   {
-    return new JoinDataSource(
+    return create(
         left,
         right,
-        StringUtils.nullToEmptyNonDruidDataString(rightPrefix),
-        JoinConditionAnalysis.forExpression(
-            Preconditions.checkNotNull(condition, "condition"),
-            StringUtils.nullToEmptyNonDruidDataString(rightPrefix),
-            macroTable
-        ),
+        rightPrefix,
+        condition,
         joinType,
-        leftFilter
+        leftFilter,
+        Suppliers.memoize(() -> Parser.parse(condition, macroTable))
     );
   }