--- a/processing/src/main/java/org/apache/druid/query/aggregation/SimpleFloatAggregatorFactory.java
+++ b/processing/src/main/java/org/apache/druid/query/aggregation/SimpleFloatAggregatorFactory.java
@@ -59,11 +59,28 @@ public abstract class SimpleFloatAggregatorFactory extends NullableNumericAggreg
       @Nullable String expression
   )
   {
+    this(
+        macroTable,
+        name,
+        fieldName,
+        expression,
+        Suppliers.memoize(() -> expression == null ? null : Parser.parse(expression, macroTable))
+    );
+  }
+
+  public SimpleFloatAggregatorFactory(
+      ExprMacroTable macroTable,
+      String name,
+      String fieldName,
+      @Nullable String expression,
+      Supplier<Expr> expressionSupplier
+  )
+  {
     this.macroTable = macroTable;
     this.name = name;
     this.fieldName = fieldName;
     this.expression = expression;
-    this.fieldExpression = Suppliers.memoize(() -> expression == null ? null : Parser.parse(expression, macroTable));
+    this.fieldExpression = expressionSupplier;
     Preconditions.checkNotNull(name, "Must have a valid, non-null aggregator name");
     Preconditions.checkArgument(
         fieldName == null ^ expression == null,