--- a/processing/src/main/java/org/apache/druid/query/aggregation/post/ExpressionPostAggregator.java
+++ b/processing/src/main/java/org/apache/druid/query/aggregation/post/ExpressionPostAggregator.java
@@ -101,26 +101,6 @@ public class ExpressionPostAggregator implements PostAggregator
       final String expression,
       @Nullable final String ordering,
       final ExprMacroTable macroTable,
-      final Supplier<Expr> parsed
-  )
-  {
-    this(
-        name,
-        expression,
-        ordering,
-        null, // in the future this will be computed by decorate
-        macroTable,
-        ImmutableMap.of(),
-        parsed,
-        Suppliers.memoize(() -> parsed.get().analyzeInputs().getRequiredBindings())
-    );
-  }
-
-  private ExpressionPostAggregator(
-      final String name,
-      final String expression,
-      @Nullable final String ordering,
-      final ExprMacroTable macroTable,
       final Map<String, Function<Object, Object>> finalizers,
       final Supplier<Expr> parsed
   )