--- a/processing/src/main/java/org/apache/druid/segment/join/JoinConditionAnalysis.java
+++ b/processing/src/main/java/org/apache/druid/segment/join/JoinConditionAnalysis.java
@@ -109,7 +109,7 @@ public class JoinConditionAnalysis
       final String condition,
       final String rightPrefix,
       final Supplier<Expr> exprSupplier
-      )
+  )
   {
     final Expr conditionExpr = exprSupplier.get();
     final List<Equality> equiConditions = new ArrayList<>();