… docs/configuration/index.md
… docs/querying/sql.md
… sql/src/main/java/org/apache/druid/sql/calcite/aggregation/builtin/SimpleSqlAggregator.java
… sql/src/main/java/org/apache/druid/sql/calcite/aggregation/builtin/SumSqlAggregator.java
… sql/src/main/java/org/apache/druid/sql/calcite/expression/DruidExpression.java
… sql/src/main/java/org/apache/druid/sql/calcite/expression/builtin/ArrayContainsOperatorConversion.java
… sql/src/main/java/org/apache/druid/sql/calcite/expression/builtin/ArrayOverlapOperatorConversion.java
… sql/src/main/java/org/apache/druid/sql/calcite/expression/builtin/BaseExpressionDimFilterOperatorConversion.java
… sql/src/main/java/org/apache/druid/sql/calcite/expression/builtin/CastOperatorConversion.java
… sql/src/main/java/org/apache/druid/sql/calcite/expression/builtin/DateTruncOperatorConversion.java
… sql/src/main/java/org/apache/druid/sql/calcite/expression/builtin/QueryLookupOperatorConversion.java
… sql/src/main/java/org/apache/druid/sql/calcite/expression/builtin/RegexpExtractOperatorConversion.java
… sql/src/main/java/org/apache/druid/sql/calcite/expression/builtin/TimeFloorOperatorConversion.java
… sql/src/main/java/org/apache/druid/sql/calcite/expression/builtin/TruncateOperatorConversion.java
… sql/src/main/java/org/apache/druid/sql/calcite/filtration/BottomUpTransform.java
… sql/src/main/java/org/apache/druid/sql/calcite/planner/CachingExprParser.java
… sql/src/main/java/org/apache/druid/sql/calcite/planner/CalcitePlannerModule.java
… sql/src/main/java/org/apache/druid/sql/calcite/planner/DruidRexExecutor.java
… sql/src/main/java/org/apache/druid/sql/calcite/planner/PlannerConfig.java
… sql/src/main/java/org/apache/druid/sql/calcite/planner/PlannerContext.java
… sql/src/main/java/org/apache/druid/sql/calcite/rel/DruidJoinQueryRel.java
… sql/src/main/java/org/apache/druid/sql/calcite/rel/DruidQuery.java
… sql/src/main/java/org/apache/druid/sql/calcite/rel/Grouping.java
… sql/src/main/java/org/apache/druid/sql/calcite/rel/Projection.java
… sql/src/main/java/org/apache/druid/sql/calcite/rel/VirtualColumnRegistry.java
… sql/src/test/java/org/apache/druid/sql/calcite/SqlPlannerExprCacheTest.java
… sql/src/test/java/org/apache/druid/sql/calcite/planner/CachingExprParserTest.java
… sql/src/test/java/org/apache/druid/sql/calcite/planner/CalcitePlannerModuleTest.java