### Description

Druid can parse the same expressions over again while it explores the search space of candidate plans. This PR adds caching parsed expressions per query. This could also be useful when the search space is large. Caching is disabled by default.

`SqlBenchmark.planSql()` results after applying this and https://github.com/apache/druid/pull/11041. Only the query 19 (the giant union query) shows a noticeable improvement (total ~25%, ~7% for only caching). However, in my testing after I quadrupled the query 19, the gain in planning time was smaller (~11%) than with the original query. This was because of too many candidate plans to explore. I think parallelizing query planning can help in this case, but didn't investigate that idea much because it should be done in Calcite.

```
master

Benchmark             (query)  (rowsPerSegment)  (vectorize)  Mode  Cnt    Score   Error  Units
SqlBenchmark.planSql        0                 5        false  avgt   15    0.563 ± 0.005  ms/op
SqlBenchmark.planSql        1                 5        false  avgt   15    0.603 ± 0.004  ms/op
SqlBenchmark.planSql        2                 5        false  avgt   15    0.662 ± 0.007  ms/op
SqlBenchmark.planSql        3                 5        false  avgt   15    0.808 ± 0.009  ms/op
SqlBenchmark.planSql        4                 5        false  avgt   15    1.156 ± 0.014  ms/op
SqlBenchmark.planSql        5                 5        false  avgt   15    1.291 ± 0.011  ms/op
SqlBenchmark.planSql        6                 5        false  avgt   15    1.504 ± 0.011  ms/op
SqlBenchmark.planSql        7                 5        false  avgt   15    1.270 ± 0.013  ms/op
SqlBenchmark.planSql        8                 5        false  avgt   15    2.044 ± 0.013  ms/op
SqlBenchmark.planSql        9                 5        false  avgt   15    1.982 ± 0.022  ms/op
SqlBenchmark.planSql       10                 5        false  avgt   15    0.707 ± 0.005  ms/op
SqlBenchmark.planSql       11                 5        false  avgt   15    0.741 ± 0.002  ms/op
SqlBenchmark.planSql       12                 5        false  avgt   15    0.624 ± 0.006  ms/op
SqlBenchmark.planSql       13                 5        false  avgt   15    0.874 ± 0.007  ms/op
SqlBenchmark.planSql       14                 5        false  avgt   15    0.980 ± 0.008  ms/op
SqlBenchmark.planSql       15                 5        false  avgt   15    0.628 ± 0.007  ms/op
SqlBenchmark.planSql       16                 5        false  avgt   15    0.745 ± 0.005  ms/op
SqlBenchmark.planSql       17                 5        false  avgt   15    0.956 ± 0.009  ms/op
SqlBenchmark.planSql       18                 5        false  avgt   15    1.078 ± 0.011  ms/op
SqlBenchmark.planSql       19                 5        false  avgt   15  129.221 ± 0.725  ms/op

patch

Benchmark             (query)  (rowsPerSegment)  (vectorize)  Mode  Cnt   Score   Error  Units
SqlBenchmark.planSql        0                 5        false  avgt   15   0.541 ± 0.003  ms/op
SqlBenchmark.planSql        1                 5        false  avgt   15   0.608 ± 0.004  ms/op
SqlBenchmark.planSql        2                 5        false  avgt   15   0.650 ± 0.003  ms/op
SqlBenchmark.planSql        3                 5        false  avgt   15   0.700 ± 0.008  ms/op
SqlBenchmark.planSql        4                 5        false  avgt   15   1.131 ± 0.007  ms/op
SqlBenchmark.planSql        5                 5        false  avgt   15   1.121 ± 0.004  ms/op
SqlBenchmark.planSql        6                 5        false  avgt   15   1.315 ± 0.015  ms/op
SqlBenchmark.planSql        7                 5        false  avgt   15   1.166 ± 0.011  ms/op
SqlBenchmark.planSql        8                 5        false  avgt   15   1.769 ± 0.014  ms/op
SqlBenchmark.planSql        9                 5        false  avgt   15   1.587 ± 0.017  ms/op
SqlBenchmark.planSql       10                 5        false  avgt   15   0.710 ± 0.003  ms/op
SqlBenchmark.planSql       11                 5        false  avgt   15   0.732 ± 0.005  ms/op
SqlBenchmark.planSql       12                 5        false  avgt   15   0.584 ± 0.005  ms/op
SqlBenchmark.planSql       13                 5        false  avgt   15   0.801 ± 0.005  ms/op
SqlBenchmark.planSql       14                 5        false  avgt   15   0.894 ± 0.005  ms/op
SqlBenchmark.planSql       15                 5        false  avgt   15   0.583 ± 0.007  ms/op
SqlBenchmark.planSql       16                 5        false  avgt   15   0.691 ± 0.003  ms/op
SqlBenchmark.planSql       17                 5        false  avgt   15   0.820 ± 0.005  ms/op
SqlBenchmark.planSql       18                 5        false  avgt   15   0.937 ± 0.007  ms/op
SqlBenchmark.planSql       19                 5        false  avgt   15  97.453 ± 0.540  ms/op
```

<hr>

##### Key changed/added classes in this PR
 * `CachingExprParser`

<hr>

<!-- Check the items by putting "x" in the brackets for the done things. Not all of these items apply to every PR. Remove the items which are not done or not relevant to the PR. None of the items from the checklist below are strictly necessary, but it would be very helpful if you at least self-review the PR. -->

This PR has:
- [x] been self-reviewed.
   - [ ] using the [concurrency checklist](https://github.com/apache/druid/blob/master/dev/code-review/concurrency.md) (Remove this item if the PR doesn't have any relation to concurrency.)
- [x] added documentation for new or modified features or behaviors.
- [x] added Javadocs for most classes and all non-trivial methods. Linked related entities via Javadoc links.
- [ ] added or updated version, license, or notice information in [licenses.yaml](https://github.com/apache/druid/blob/master/dev/license.md)
- [ ] added comments explaining the "why" and the intent of the code wherever would not be obvious for an unfamiliar reader.
- [x] added unit tests or modified existing tests to cover new code paths, ensuring the threshold for [code coverage](https://github.com/apache/druid/blob/master/dev/code-review/code-coverage.md) is met.
- [ ] added integration tests.
- [ ] been tested in a test Druid cluster.