@jihoonson 2021-03-28
I created https://github.com/apache/druid/pull/11041 so that we can focus on caching expressions in this PR.